Fix issues related to merge commit
